Compare Amir's similarities

+++ tests/numpy_game.py
@@ -14,7 +14,7 @@ def file012_to_numpy(input_file_path, raw_file=None):
     if split_individuals[-1] == '':  # we throw empty line at the end of the file
         split_individuals = split_individuals[:-1]
     split_sites = [individual.split('\t') for individual in split_individuals]
-    arr = np.array(split_sites, dtype=np.int8)
+    arr = np.array(split_sites)
     if np.any(arr[:, 0] > 2):
         arr = arr[:, 1:]  # First column is individual number.
     return arr
